+++ internal/cli/root.go
@@ -3,6 +3,7 @@ package cli
 import (
 	"fmt"
 	"os"
+	"runtime/debug"
 
 	"github.com/spf13/cobra"
 )
@@ -41,10 +42,58 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print version information",
 	Run: func(cmd *cobra.Command, args []string) {
+		info := effectiveBuildInfo(readVCSBuildSettings())
 		fmt.Printf("haft %s\n", Version)
-		fmt.Printf("  commit:  %s\n", Commit)
-		fmt.Printf("  built:   %s\n", BuildDate)
+		fmt.Printf("  commit:  %s\n", info.Commit)
+		fmt.Printf("  built:   %s\n", info.BuildDate)
+		if info.SourceTime != "" {
+			fmt.Printf("  source:  %s\n", info.SourceTime)
+		}
+		if info.Modified {
+			fmt.Printf("  modified: true\n")
+		}
 	},
+}
+
+type buildInfo struct {
+	Commit     string
+	BuildDate  string
+	SourceTime string
+	Modified   bool
+}
+
+func effectiveBuildInfo(settings map[string]string) buildInfo {
+	info := buildInfo{
+		Commit:    Commit,
+		BuildDate: BuildDate,
+	}
+	if info.Commit == "" || info.Commit == "none" {
+		if revision := settings["vcs.revision"]; revision != "" {
+			info.Commit = revision
+		}
+	}
+	if info.Commit == "" {
+		info.Commit = "none"
+	}
+	if info.BuildDate == "" {
+		info.BuildDate = "unknown"
+	}
+	info.SourceTime = settings["vcs.time"]
+	info.Modified = settings["vcs.modified"] == "true"
+	return info
+}
+
+func readVCSBuildSettings() map[string]string {
+	info, ok := debug.ReadBuildInfo()
+	if !ok {
+		return nil
+	}
+
+	settings := make(map[string]string, len(info.Settings))
+	for _, setting := range info.Settings {
+		settings[setting.Key] = setting.Value
+	}
+	return settings
 }
 
 func Execute() {